requirements.txt
gradio==6.0.2
huggingface_hub
python-dotenv